refactor(test-kernel): share TS parse and module-specifier scanning across ratchets
Nine architecture ratchets each hand-rolled "parse a TypeScript file and
collect the module specifiers it loads". `src/test-kernel/support/repoScan.ts`
already declares itself the home for exactly this plumbing and already owns the
file-walking half, so the parse call and the plain specifier scan move there.

The collectors had drifted, and the drift is a live hole:
agentRuntimeProgressEventsBoundary had no `ts.isImportTypeNode` branch, so a
second production file writing `typeof import('.../cliNdjsonProgressEvents')`
passed that guard untouched while two sibling ratchets caught the identical
form. Fifteen production `typeof import(...)` sites already exist, including
relative ones under src/tools/. Adopting the superset collector closes it.

The `.tsx ? ScriptKind.TSX : ScriptKind.TS` ternary was itself reinvented:
ts.createSourceFile derives ScriptKind from the file name when the argument is
omitted, so six copies delete outright rather than relocating. The two
genuinely richer collectors stay put: subsystemEdgeRatchet returns an EdgeKind
alongside the specifier, sharedSchemasDeepImportRatchet returns per-binding
type-space, and supportMockImportOrder must stay non-recursive over
`sourceFile.statements`. All three take the shared parse and keep their own
extraction; sharedSchemasDeepImportRatchet keeps its documented
`setParentNodes: false` perf choice as an option.

No ratchet baseline is regenerated or widened.
