+++ src/test-kernel/architecture/hostedAuthImportBoundary.vitest.ts
@@ -3,10 +3,15 @@ import { readFileSync } from 'node:fs';
 import { dirname, resolve } from 'node:path';
 
 // Third-party imports
-import ts from 'typescript';
 import { describe, expect, it } from 'vitest';
 
-import { REPO_ROOT, sourceFilesUnder, toRepoPath } from '../support/repoScan';
+import {
+  collectModuleSpecifiers,
+  parseSourceFile,
+  REPO_ROOT,
+  sourceFilesUnder,
+  toRepoPath,
+} from '../support/repoScan';
 
 const SCAN_ROOTS = [
   'src/model',
@@ -22,72 +27,6 @@ const PERMITTED_THIRD_PARTY_OAUTH_PATHS = [
   'src/auth/codex',
   'src/auth/xai',
 ] as const;
-
-function parseSourceFile(file: string, source: string): ts.SourceFile {
-  return ts.createSourceFile(
-    file,
-    source,
-    ts.ScriptTarget.Latest,
-    true,
-    file.endsWith('.tsx') ? ts.ScriptKind.TSX : ts.ScriptKind.TS,
-  );
-}
-
-function stringLiteralText(node: ts.Node): string | null {
-  return ts.isStringLiteralLike(node) ? node.text : null;
-}
-
-function unwrapExpression(node: ts.Expression): ts.Expression {
-  while (
-    ts.isParenthesizedExpression(node) ||
-    ts.isAsExpression(node) ||
-    ts.isTypeAssertionExpression(node) ||
-    ts.isSatisfiesExpression(node)
-  ) {
-    node = node.expression;
-  }
-  return node;
-}
-
-function moduleSpecifier(node: ts.Node): string | null {
-  if (ts.isImportDeclaration(node) || ts.isExportDeclaration(node)) {
-    return node.moduleSpecifier == null
-      ? null
-      : stringLiteralText(node.moduleSpecifier);
-  }
-  if (ts.isImportEqualsDeclaration(node)) {
-    const reference = node.moduleReference;
-    return ts.isExternalModuleReference(reference)
-      ? stringLiteralText(reference.expression)
-      : null;
-  }
-  if (ts.isImportTypeNode(node) && ts.isLiteralTypeNode(node.argument)) {
-    return stringLiteralText(node.argument.literal);
-  }
-  if (ts.isCallExpression(node)) {
-    const [argument] = node.arguments;
-    const loadsModule =
-      node.expression.kind === ts.SyntaxKind.ImportKeyword ||
-      (ts.isIdentifier(node.expression) && node.expression.text === 'require');
-    return loadsModule && argument != null
-      ? stringLiteralText(unwrapExpression(argument))
-      : null;
-  }
-  return null;
-}
-
-function collectModuleSpecifiers(sourceFile: ts.SourceFile): string[] {
-  const specifiers: string[] = [];
-
-  const visit = (node: ts.Node): void => {
-    const specifier = moduleSpecifier(node);
-    if (specifier != null) specifiers.push(specifier);
-    ts.forEachChild(node, visit);
-  };
-
-  visit(sourceFile);
-  return specifiers;
-}
 
 function authRoot(specifier: string): string | null {
   if (specifier === '@auth') return '@auth';
@@ -130,9 +69,9 @@ function isForbiddenAuthImport(file: string, specifier: string): boolean {
 }
 
 function findForbiddenAuthImports(file: string, source: string): string[] {
-  return collectModuleSpecifiers(parseSourceFile(file, source)).filter(
-    (specifier) => isForbiddenAuthImport(file, specifier),
-  );
+  return collectModuleSpecifiers(
+    parseSourceFile(file, { text: source }),
+  ).filter((specifier) => isForbiddenAuthImport(file, specifier));
 }
 
 describe('hosted auth import boundary', () => {
@@ -188,6 +127,24 @@ describe('hosted auth import boundary', () => {
         `,
       ),
     ).toEqual(['../../auth/SupabaseClient', '../../auth/sessionToken']);
+  });
+
+  // Regression: the collector this suite shares with the other ratchets grew
+  // its `ts.isImportTypeNode` branch unevenly — agentRuntimeProgressEventsBoundary
+  // had none, so a `typeof import(…)` reference slipped past a
+  // single-permitted-importer guard while two siblings caught the same form.
+  // Nothing else in the repo exercises a type position, so pin it here.
+  it('rejects typeof-import type positions, aliased and relative', () => {
+    expect(
+      findForbiddenAuthImports(
+        resolve(REPO_ROOT, 'src/agent/runtime/fixture.ts'),
+        `
+          type Client = typeof import('@auth/SupabaseClient');
+          type Session = import('../../auth/sessionToken').Session;
+          type Codex = typeof import('@auth/codex/oauth');
+        `,
+      ),
+    ).toEqual(['@auth/SupabaseClient', '../../auth/sessionToken']);
   });
 
   it('rejects asserted module arguments and dynamic imports with options', () => {
